third_party/WebKit/LayoutTests/platform/linux/plugins/npp-set-window-called-during-destruction-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: undefined is not a function

This tests whether a NPP_SetWindow is called during plugin destruction.
